+++ config/placeholders.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    'image' => '/public/vendor/images/placeholders/image.jpg',
+    'logo' => '/public/vendor/images/placeholders/logo.png',
+    'cover' => '/public/vendor/images/placeholders/cover.jpg',
+    'dashboard_cover' => '/public/vendor/images/placeholders/dashboard_cover.jpg',
+];