+++ resources/views/admin/templates/index.blade.php
@@ -1,0 +1,58 @@
+@extends('pakka::admin.default')
+
+@section('page-header')
+    @if(Session::has('module_name'))
+		{{ Session::get('module_name') }}
+	@endif
+	 
+	<small>{{ trans('pakka::app.manage') }}</small>
+@endsection
+
+@section('content')
+
+    <div class="mB-20">
+        <a href="{{ route(config('pakka.prefix.admin'). '.templates.create') }}" class="btn btn-info">
+            {{ trans('pakka::app.add_button') }}
+        </a>
+    </div>
+
+
+    <div class="table-container bgc-white bd bdrs-3 p-20 mB-20">
+        <table id="dataTable" class="table table-list table-striped" cellspacing="0" width="100%"> <!-- table-bordered -->
+            <thead>
+                <tr>
+                    <th>{{ trans('pakka::app.name') }}</th>
+                    <th>{{ trans('pakka::app.download') }}</th>
+                    <th></th>
+                </tr>
+            </thead>
+            
+            <tbody>
+                @foreach ($templates as $template)
+                    <tr>
+                        <td>{!! $template->name !!}</td>
+                        <td><a href="{{ route(config('pakka.prefix.admin'). '.templates.download', $template->id) }}" class="btn btn-primary btn-sm">{{ trans('pakka::app.download') }}</a></td>
+                        <td>
+                            <ul class="list-inline">
+                                <li class="list-inline-item">
+                                    {!! Form::open([
+                                        'class'=>'delete',
+                                        'url'  => route(config('pakka.prefix.admin'). '.templates.destroy', $template->id), 
+                                        'method' => 'DELETE',
+                                        ]) 
+                                    !!}
+
+                                        <button class="btn btn-danger btn-sm" title="{{ trans('pakka::app.delete_title') }}"><i class="ti-trash"></i></button>
+                                        
+                                    {!! Form::close() !!}
+                                </li>
+                            </ul>
+                        </td>
+                    </tr>
+                @endforeach
+            </tbody>
+        
+        </table>
+    </div>
+
+@endsection